Add files via upload
In this Demo
This is a activity that depicts the use of Loops like while, if and else. In this the user writes a number and then a random number is generated later the user has to guess what number it might be. It also shows as to how many attempts the user has made to guess the number.